+++ App.js
@@ -13,6 +13,7 @@ import {backgroundThemeColor} from './styles/globalStyles';
 import HomePage from './screens/HomePage';
 import Login from './screens/Login';
 import Signup from './screens/Signup';
+import ContactUs from './screens/ContactUs';
 
 const App = () => {
   const isDarkMode = useColorScheme() === 'dark';
@@ -32,6 +33,7 @@ const App = () => {
       {/* <HomePage /> */}
       {/* <Login /> */}
       {/* <Signup /> */}
+      <ContactUs />
     </View>
   );
 };
